Update event index and routes in server
