+++ app/controllers/server.rb
@@ -18,7 +18,6 @@ module TrafficSpy
       end
     end
 
-    # event index
     get '/sources/:identifier/events' do
       @source = Source.find_by!(identifier: params[:identifier])
       events_by_source = @source.payloads.map {|payload| payload.event }
@@ -29,9 +28,7 @@ module TrafficSpy
     get '/sources/:indentifier/events/:EVENTNAME' do    
       #add sad path page if event is not defined
       #link back to events index page
-      erb :app_event_details
-
-      
+      erb :event_details
     end
 
     post '/sources/:identifier/data' do |identifier|
