--- app/views/app_event_details.erb
@@ -1,21 +1,0 @@
-
-<h2>Application Event Details</h2>
-
-	<h4>Hour by Hour Breakdown of when the event was received:</h4>
-
-		<ol>
-		<!--#name of event
-	#0 to 23 hours with a tally-->
-			<li></li>
-			<li></li>
-			<li></li>
-		</ol>
-
-
-</div>
-
-
-<div id='events-details-overall'>
-	<h4>How many times the event was received overall:</h4>
-	<p>this is a number</p>
-</div>
