+++ app/views/event_index.erb
@@ -5,7 +5,7 @@
 
 	<ol id="events">
 		<%@events.each do |(event, count)| %>
-		<li> <a href="/sources/<%= @source.identifier%>/events/<%= event.name %>"><%=event.name%></a></li>
+		<li> <a href="/sources/<%= @source.identifier%>/events"><%=event.name%></a></li>
 		<% end %>
 
 	</ol>
